Eliminate global usage of torch.set_default_dtype in test_autograd
[ghstack-poisoned]

+++ torch/testing/_internal/common_methods_invocations.py
@@ -4841,10 +4841,10 @@ spectral_funcs = [op for op in op_db if isinstance(op, SpectralFuncInfo)]
 sparse_unary_ufuncs = [op for op in op_db if isinstance(op, UnaryUfuncInfo) and op.supports_sparse is True]
 shape_funcs = [op for op in op_db if isinstance(op, ShapeFuncInfo)]
 
-def index_variable(shape, max_indices, device=torch.device('cpu')):
+def index_variable(shape, max_indices, dtype=torch.double, device=torch.device('cpu')):
     if not isinstance(shape, tuple):
         shape = (shape,)
-    index = torch.rand(*shape, device=device).mul_(max_indices).floor_().long()
+    index = torch.rand(*shape, dtype=dtype, device=device).mul_(max_indices).floor_().long()
     return index
 
 
@@ -5301,7 +5301,7 @@ def create_input(call_args, requires_grad=True, non_contiguous=False, call_kwarg
         # double check casting
         elif isinstance(arg, non_differentiable):
             if isinstance(arg.tensor, torch.Tensor):
-                return maybe_non_contig(arg.tensor.to(device=device))
+                return maybe_non_contig(arg.tensor.to(dtype=dtype, device=device))
             return maybe_non_contig(arg.tensor.to(device=device))
         elif isinstance(arg, torch.Tensor):
             if arg.dtype == torch.float:
